requirements/project.txt
Mezzanine==3.0.9
Django==1.6.2
requests==2.0.1
feedparser==5.1.3
django-taggit==0.11.2
South==0.8.4
Sphinx==1.2
BeautifulSoup==3.2.1
lxml==3.3.3
django-debug-toolbar==1.0
django-extensions==1.2.5
django-filter==0.7
djangorestframework==2.3.11
pathod==0.10
flake8==2.1.0
django-mustachejs==0.8.3
splinter==0.6.0
behave==1.2.3
mock==1.0.1
Pillow==2.1.0
readability-lxml==0.3.0.2
django-polymorphic==0.5.4
-e git://github.com/dgreisen-cfpb/django-behave.git@my_fixes#egg=django-behave
celery==3.1
